// src/trade/interfaces/dvla-vehicle.interface.ts
export interface DvlaVehicle {
  dvlaVehicleId: string;
  registration: string;
  firstRegistrationDate: string;
  makeName: string;
  modelName: string;
  primaryColour: string;
  secondaryColour: string;
  fuelType: string;
  euClassification: string;
  bodyTypeCode: string;
  hasLinkedDvsaVehicle: boolean;
  motTestDueDate: string;
}